chore(lint): выправить clippy по всему workspace и поставить lint-гейт в CI
Долг прежних этапов: ~150 предупреждений clippy, которые CI не проверял.

Настоящие правки, а не подавления:
- ToolError::NeedsApproval носит Box<ApprovalRequired> — раньше payload
  approval'а (полный вход инструмента и preview) клал ~250 байт на каждый
  Result<_, ToolError> в крейте, включая happy path;
- approval-API permissions принимает CallIdentity вместо пяти позиционных
  аргументов: approval сверяется с вызовом целиком, а перепутанные местами
  tool_name/scope дали бы «совпадение» там, где его нет;
- tests/memory_api_contract.rs и tests/routing_runtime.rs удалены: они
  подключали модуль из src через #[path] и гоняли те же тесты второй раз.
  Вместе с ними ушла cfg(test)-зеркальная копия routing_policy внутри
  routing_runtime и раздвоение импортов в lib.rs model-gateway;
- SQL-строки в receipts получили type-алиасы вместо 11–13-местных кортежей;
- drop(runtime) в пяти местах не был нужен: NLL закрывает заимствование сам;
- RoutingRuntime::mode() читает поле, которое до сих пор только писалось;
- остальное — мелочь, снятая cargo clippy --fix.

Осознанные allow с обоснованием: bounded-контракты supervisor'а, чья
поверхность пока вызывается только собственными тестами, и функции, у которых
список аргументов повторяет колонки строки SQLite или поля трассы.

В CI добавлен шаг cargo clippy --workspace --all-targets -- -D warnings.

Тесты всех крейтов, receipt-key-lifecycle, eval-gate и security-eval-gate
проходят. Число тестов evohime-core 369 -> 363: ушли шесть дублей
memory_api::domain::tests, идентичных memory_domain::tests.

+++ crates/desktop-ipc/src/windows_security.rs
@@ -58,7 +58,7 @@ impl PipeSecurity {
         let created = unsafe {
             ConvertStringSecurityDescriptorToSecurityDescriptorW(
                 wide.as_ptr(),
-                SDDL_REVISION_1 as u32,
+                SDDL_REVISION_1,
                 &mut descriptor,
                 std::ptr::null_mut(),
             )
